Fix deprecated warnings
