+++ ic-log/src/settings.rs
@@ -1,3 +1,6 @@
+// This is not to generate warning for deriving traits on a deprecated structure.
+#![allow(deprecated)]
+
 use candid::{CandidType, Principal};
 use serde::Deserialize;
 
@@ -5,6 +8,26 @@ use crate::did::{LogCanisterSettings, LoggerAcl, LoggerPermission};
 
 const DEFAULT_IN_MEMORY_RECORDS: usize = 1024;
 const DEFAULT_MAX_RECORD_LENGTH: usize = 1024;
+
+/// Log settings to initialize the logger
+#[derive(Default, Debug, Clone, CandidType, Deserialize)]
+#[deprecated(
+    since = "0.19.0",
+    note = "newer api uses `LogSettingsV2`, use this type for older canister calls only"
+)]
+pub struct LogSettings {
+    /// Enable logging to console (`ic::print` when running in IC)
+    pub enable_console: bool,
+    /// Number of records to be stored in the circular memory buffer.
+    /// If None - storing records will be disable.
+    /// If Some - should be power of two.
+    pub in_memory_records: Option<usize>,
+    /// Log configuration as combination of filters. By default the logger is OFF.
+    /// Example of valid configurations:
+    /// - info
+    /// - debug,crate1::mod1=error,crate1::mod2,crate2=debug
+    pub log_filter: Option<String>,
+}
 
 /// Logger settings.
 ///
